0.0.10.0 Release
This mod is pushed early as this mod is broken by a significant modification that the base PMDO accepted. There may be some bugs, and you may need to file bug reports. When new issues are identified and development advancements are made, the next bugfix may be released in a batch.